T.P.Scripting: add CustomComponents, and change...
...the ScriptingEngine fields.  Instead of separate fields for
custom readers and writers, we now have a single function
that loads any number of "components" from a script: these
may be custom readers, custom writers, templates for writers,
or extension configs.  (Note: it's possible to have a custom
reader and a custom writer for a format together in the same file.)

This is primarily motivated by the need for a smoother check
for custom templates and extensions for `-D` and `--list-extensions`,
but it also seems more elegant.

This is a partial implementation of #8417.

What remains are the changes in pandoc-lua-engine.
We need to replace T.P.Lua.Writer and T.P.Lua.Reader with
a single module T.P.Lua.Custom that implements
`customLoadScript`.  I imagine most of the code from
the existing two modules can be reused here.